04-service | fix a build and new config splitted on audit and main
