+++ configs/config.yaml
@@ -11,7 +11,7 @@ audit:
 database:
     db_core: postgres
     db_user: postgres
-    db_name: test
     db_password: postgres
+    db_name: test
     db_port: 5432
     db_connections_pool: 10
